perf(react-query): share shop-settings cache for admin branding chrome
- Extract fetchShopSettings + shopSettingsQueryKey for GET /shop-settings
- ShopSettingsPage and useAdminShopBranding use same queryKey/queryFn
- useAdminShopBranding selects appearance only (no duplicate network)
